Added initial vocab list controller tests and move fluent validators around.
