+++ GermanVocabApp.Api.FluentValidation.Tests.Unit/ValidatorTests/FluentModifierTests.cs
@@ -1,19 +1,17 @@
-using FluentValidation.TestHelper;
+﻿using FluentValidation.TestHelper;
 using GermanVocabApp.Api.FluentValidation.Validators;
 using GermanVocabApp.Shared.Data;
-using Moq;
 
-namespace GermanVocabApp.Api.FluentValidation.Tests.Unit;
+namespace GermanVocabApp.Api.FluentValidation.Tests.Unit.ValidatorTests;
 
-public class FluentNounValidatorTests : FluentWordValidatorTests<FluentNounValidator>
+public class FluentModifierValidatorTests : FluentWordValidatorTests<FluentModifierValidator>
 {
-    protected override FluentNounValidator CreateValidator()
+    protected override FluentModifierValidator CreateValidator()
     {
-        return new FluentNounValidator();
+        return new FluentModifierValidator();
     }
 
-
-    #region IsWeakmasculineNoun
+    #region IsWeakMasculineNoun
     [Theory]
     [InlineData(true)]
     [InlineData(false)]
@@ -21,16 +19,15 @@ public class FluentNounValidatorTests : FluentWordValidatorTests<FluentNounValid
     {
         Mock.Setup(r => r.IsWeakMasculineNoun).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.IsWeakMasculineNoun);
+        result.ShouldHaveValidationErrorFor(request => request.IsWeakMasculineNoun);
     }
 
-    [Theory]
-    [InlineData(null)]
-    public void IsWeakMasculineNoun_ShouldHaveValidationError_WhenNull(bool? value)
+    [Fact]
+    public void IsWeakMasculineNoun_ShouldNotHaveValidationError_WhenNull()
     {
-        Mock.Setup(r => r.IsWeakMasculineNoun).Returns(value);
+        Mock.Setup(r => r.IsWeakMasculineNoun).Returns(() => null);
         var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.IsWeakMasculineNoun);
+        result.ShouldNotHaveValidationErrorFor(request => request.IsWeakMasculineNoun);
     }
     #endregion
 
@@ -96,7 +93,7 @@ public class FluentNounValidatorTests : FluentWordValidatorTests<FluentNounValid
     }
     #endregion
 
-    #region ThirdPersonPresent
+    #region ThidPersonPresent
     [Fact]
     public void ThirdPersonPresent_ShouldNotHaveValidationError_WhenNull()
     {
@@ -120,7 +117,7 @@ public class FluentNounValidatorTests : FluentWordValidatorTests<FluentNounValid
     [Fact]
     public void ThirdPersonImperfect_ShouldNotHaveValidationError_WhenNull()
     {
-        Mock.Setup(r => r.ThirdPersonImperfect).Returns((string)null);
+        Mock.Setup(r => r.ThirdPersonImperfect).Returns(() => null);
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldNotHaveValidationErrorFor(request => request.ThirdPersonImperfect);
     }
@@ -177,23 +174,45 @@ public class FluentNounValidatorTests : FluentWordValidatorTests<FluentNounValid
     #endregion
 
     #region Gender
-    [Fact]
-    public void Gender_ShouldHaveValidationError_WhenNull()
-    {
-        Mock.Setup(r => r.Gender).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.Gender);
-    }
-
     [Theory]
     [InlineData(Gender.Masculine)]
     [InlineData(Gender.Feminine)]
     [InlineData(Gender.Neuter)]
-    public void Gender_ShouldNotHaveValidationError_WhenNotNull(Gender value)
+    public void Gender_ShouldHaveValidationError_WhenNotNull(Gender value)
     {
         Mock.Setup(r => r.Gender).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
+        result.ShouldHaveValidationErrorFor(request => request.Gender);
+    }
+
+    [Fact]
+    public void Gender_ShouldNotHaveValidationError_WhenNull()
+    {
+        Mock.Setup(r => r.Gender).Returns(() => null);
+        var result = Validator.TestValidate(Mock.Object);
         result.ShouldNotHaveValidationErrorFor(request => request.Gender);
+    }
+    #endregion
+
+    #region Plural
+    [Theory]
+    [InlineData(StringData.Empty)]
+    [InlineData(StringData.Whitespace)]
+    [InlineData("abc")]
+    [InlineData(StringData.CharString25)]
+    public void Plural_ShouldHaveValidationError_WhenNotNull(string? value)
+    {
+        Mock.Setup(r => r.Plural).Returns(value);
+        var result = Validator.TestValidate(Mock.Object);
+        result.ShouldHaveValidationErrorFor(request => request.Plural);
+    }
+
+    [Fact]
+    public void Plural_ShouldNotHaveValidationError_WhenNull()
+    {
+        Mock.Setup(r => r.Plural).Returns(() => null);
+        var result = Validator.TestValidate(Mock.Object);
+        result.ShouldNotHaveValidationErrorFor(request => request.Plural);
     }
     #endregion
 
@@ -202,199 +221,107 @@ public class FluentNounValidatorTests : FluentWordValidatorTests<FluentNounValid
     [InlineData(StringData.Empty)]
     [InlineData(StringData.CharString1)]
     [InlineData(StringData.CharString26)]
-    public void Preposition_ShouldHaveValidationError_WhenInvalidLength(string? value)
+    public void Preposition_ShouldHaveValidationError_WhenNotNull(string? value)
     {
         Mock.Setup(r => r.Preposition).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldHaveValidationErrorFor(request => request.Preposition);
     }
 
-    [Theory]
-    [InlineData(null)]
-    [InlineData("abc")]
-    [InlineData(StringData.CharString25)]
-    public void Preposition_ShouldNotHaveValidationError_WhenNull_OrValidLength(string? value)
+    [Fact]
+    public void Preposition_ShouldNotHaveValidationError_WhenNull()
     {
-        Mock.Setup(r => r.Preposition).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldNotHaveValidationErrorFor(request => request.Preposition);
     }
     #endregion
 
     #region PrepositionCase
-    [Fact]
-    public void PrepositionCase_ShouldNotHaveValidationError_WhenNull_AndPrepositionCaseNull()
+    [Theory]
+    [InlineData(Case.Nominative)]
+    [InlineData(Case.Accusative)]
+    [InlineData(Case.Dative)]
+    [InlineData(Case.Genetive)]
+    public void PrepositionCase_ShouldHaveValidationError_WhenNotNull(Case value)
     {
-        Mock.Setup(r => r.Preposition).Returns(() => null);
-        Mock.Setup(r => r.PrepositionCase).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.PrepositionCase);
-    }
-
-    [Fact]
-    public void PrepositionCase_ShouldHaveValidationError_WhenNotNull_AndPrepositionNull()
-    {
-        Mock.Setup(r => r.Preposition).Returns(() => null);
-        Mock.Setup(r => r.PrepositionCase).Returns(() => It.IsAny<Case>());
+        Mock.Setup(r => r.PrepositionCase).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldHaveValidationErrorFor(request => request.PrepositionCase);
     }
 
-    [Theory]
-    [InlineData(StringData.Whitespace)]
-    [InlineData(StringData.Empty)]
-    [InlineData(StringData.CharString1)]
-    public void PrepositionCase_ShouldHaveValidationError_WhenNull_AndPrepositionNotNull(string? prepositionValue)
+    [Fact]
+    public void PrepositionCase_ShouldNotHaveValidationError_WhenNull()
     {
-        Mock.Setup(r => r.Preposition).Returns(() => prepositionValue);
-        Mock.Setup(r => r.PrepositionCase).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.PrepositionCase);
-    }
-
-    [Theory]
-    [InlineData(StringData.Whitespace)]
-    [InlineData(StringData.Empty)]
-    [InlineData(StringData.CharString1)]
-    public void PrepositionCase_ShouldNotHaveValidationError_WhenNotNull_AndPrepositionNotNull(string? prepositionValue)
-    {
-        Mock.Setup(r => r.Preposition).Returns(prepositionValue);
-        Mock.Setup(r => r.PrepositionCase).Returns(() => It.IsAny<Case>());
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldNotHaveValidationErrorFor(request => request.PrepositionCase);
-    }
-    #endregion
-
-    #region Plural
-    [Theory]
-    [InlineData(StringData.Empty)]
-    [InlineData(StringData.CharString1)]
-    [InlineData(StringData.CharString2)]
-    [InlineData(StringData.CharString26)]
-    public void Plural_ShouldHaveValidationError_WhenInvalidLength(string value)
-    {
-        Mock.Setup(r => r.Plural).Returns(value);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.Plural);
-    }
-
-    [Theory]
-    [InlineData(null)]
-    [InlineData("abc")]
-    [InlineData(StringData.CharString25)]
-    public void Plural_ShouldNotHaveValidationError_WhenNull_OrValidLength(string? value)
-    {
-        Mock.Setup(r => r.Plural).Returns(value);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.Plural);
-    }
-
-    [Fact]
-    public void Plural_ShouldHaveValidationError_WhenNull_AndAlwaysPlural()
-    {
-        Mock.Setup(r => r.FixedPlurality).Returns(FixedPlurality.Plural);
-        Mock.Setup(r => r.Plural).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.Plural);
-    }
-
-    [Theory]
-    [InlineData("abc")]
-    public void Plural_ShouldNotHaveValidationError_WhenNotNull_AndAlwaysPlural(string? value)
-    {
-        Mock.Setup(r => r.FixedPlurality).Returns(FixedPlurality.Plural);
-        Mock.Setup(r => r.Plural).Returns(value);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.Plural);
-    }
-
-    [Fact]
-    public void Plural_ShouldNotHaveValidationError_WhenNull_AndAlwaysSingular()
-    {
-        Mock.Setup(r => r.FixedPlurality).Returns(FixedPlurality.Singular);
-        Mock.Setup(r => r.Plural).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.Plural);
-    }
-
-    [Theory]
-    [InlineData(StringData.Whitespace)]
-    [InlineData(StringData.Empty)]
-    [InlineData(StringData.CharString1)]
-    public void Plural_ShouldHaveValidationError_WhenNotNull_AndAlwaysSingular(string? value)
-    {
-        Mock.Setup(r => r.FixedPlurality).Returns(FixedPlurality.Singular);
-        Mock.Setup(r => r.Plural).Returns(value);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.Plural);
-    }
-
-    [Fact]
-    public void Plural_ShouldNotHaveValidationError_WhenNoFixedPlurality()
-    {
-        Mock.Setup(r => r.FixedPlurality).Returns(FixedPlurality.Singular);
-        Mock.Setup(r => r.Plural).Returns(() => It.IsAny<string>());
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.Plural);
     }
     #endregion
 
     #region Comparative
-    [Fact]
-    public void Comparative_ShouldNotHaveValidationError_WhenNull()
-    {
-        Mock.Setup(r => r.Comparative).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.Comparative);
-    }
-
     [Theory]
     [InlineData(StringData.Empty)]
     [InlineData(StringData.CharString1)]
-    public void Comparative_ShouldHaveValidationError_WhenNotNull(string value)
+    [InlineData(StringData.CharString2)]
+    [InlineData(StringData.CharString101)]
+    public void Comparative_ShouldHaveValidationError_WhenInvalidLength(string? value)
     {
         Mock.Setup(r => r.Comparative).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldHaveValidationErrorFor(request => request.Comparative);
     }
+
+    [Theory]
+    [InlineData(null)]
+    [InlineData("abc")]
+    [InlineData(StringData.CharString100)]
+    public void Comparative_ShouldNotHaveValidationError_WhenNull_OrValidLength(string? value)
+    {
+        Mock.Setup(r => r.Comparative).Returns(value);
+        var result = Validator.TestValidate(Mock.Object);
+        result.ShouldNotHaveValidationErrorFor(request => request.Comparative);
+    }
     #endregion
 
     #region Superlative
-    [Fact]
-    public void Superlative_ShouldNotHaveValidationError_WhenNull()
-    {
-        Mock.Setup(r => r.Superlative).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldNotHaveValidationErrorFor(request => request.Superlative);
-    }
-
     [Theory]
     [InlineData(StringData.Empty)]
     [InlineData(StringData.CharString1)]
-    public void Superlative_ShouldHaveValidationError_WhenNotNull(string value)
+    [InlineData(StringData.CharString2)]
+    [InlineData(StringData.CharString101)]
+    public void Superlative_ShouldHaveValidationError_WhenInvalidLength(string? value)
     {
         Mock.Setup(r => r.Superlative).Returns(value);
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldHaveValidationErrorFor(request => request.Superlative);
     }
+
+    [Theory]
+    [InlineData(null)]
+    [InlineData("abc")]
+    [InlineData(StringData.CharString100)]
+    public void Superlative_ShouldNotHaveValidationError_WhenNull_OrValidLength(string? value)
+    {
+        Mock.Setup(r => r.Superlative).Returns(value);
+        var result = Validator.TestValidate(Mock.Object);
+        result.ShouldNotHaveValidationErrorFor(request => request.Superlative);
+    }
     #endregion
 
     #region FixedPlurality
-    [Fact]
-    public void FixedPlurality_ShouldHaveValidationError_WhenNull()
-    {
-        Mock.Setup(r => r.FixedPlurality).Returns(() => null);
-        var result = Validator.TestValidate(Mock.Object);
-        result.ShouldHaveValidationErrorFor(request => request.FixedPlurality);
-    }
-
     [Theory]
     [InlineData(FixedPlurality.None)]
     [InlineData(FixedPlurality.Singular)]
     [InlineData(FixedPlurality.Plural)]
-    public void FixedPlurality_ShouldNotHaveValidationError_WhenNotNull(FixedPlurality value)
+    public void FixedPlurality_ShouldHaveValidationError_WhenNotNull(FixedPlurality value)
     {
         Mock.Setup(r => r.FixedPlurality).Returns(value);
+        var result = Validator.TestValidate(Mock.Object);
+        result.ShouldHaveValidationErrorFor(request => request.FixedPlurality);
+    }
+
+    [Fact]
+    public void FixedPlurality_ShouldNotHaveValidationError_WhenNull()
+    {
         var result = Validator.TestValidate(Mock.Object);
         result.ShouldNotHaveValidationErrorFor(request => request.FixedPlurality);
     }
